EOP moved to Telescope, slight refactor.
