+++ lib/testing/fakeGpu.hpp
@@ -111,6 +111,7 @@ public:
 
     // Dummy time map implementations
     timespec to_time(uint64_t seq) const override;
+    int64_t to_time_ns(uint64_t seq) const override;
     uint64_t to_seq(timespec time) const override;
     uint64_t seq_length_nsec() const override;
     bool gps_time_enabled() const override;
