+++ lib/utils/CHIMETelescope.cpp
@@ -17,7 +17,10 @@
 REGISTER_TELESCOPE(CHIMETelescope, "CHIMETelescope");
 
 CHIMETelescope::CHIMETelescope(const kotekan::Config& config, const std::string& path) :
-    ICETelescope(config.get<std::string>(path, "log_level")) {
+    ICETelescope(path,
+                 config.get<std::string>(path, "log_level"),
+                 config.get_default<std::string>(path, "updatable_config", "")) {
+    INFO("Building CHIMETelescope");
 
     // This is always 1 for CHIME
     _num_freq_per_stream = 1;
